requirements.txt
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
PyMySQL==1.0.2
pytz==2021.1
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.4.13
Werkzeug==1.0.1
WTForms==2.3.3
